Move BlockChainTest methods related to Append() to a new partial class
[changelog skip]
