+++ Libplanet.Tests/Menees.Analyzers.Settings.xml
@@ -2,5 +2,5 @@
 <Menees.Analyzers.Settings>
   <MaxLineColumns>100</MaxLineColumns>
   <MaxMethodLines>300</MaxMethodLines>
-  <MaxFileLines>2800</MaxFileLines>
+  <MaxFileLines>2550</MaxFileLines>
 </Menees.Analyzers.Settings>
